Refactor `Month` out to *largo/__init__.py*
